requirements.txt
Flask
azure-data-tables
python-dotenv
pandas
requests
my